Split out redux from Music Lab timeline
