+++ apps/src/music/views/hooks/usePlaybackUpdate.ts
@@ -1,21 +1,20 @@
 import {useEffect, useRef} from 'react';
 
-import {useAppSelector} from '@cdo/apps/util/reduxHooks';
-
 const UPDATE_RATE = 1000 / 30; // 30 times per second
 
 /**
  * A hook for performing updates during playback.
+ * @param isPlaying whether playback is in progress.
  * @param doUpdate function to run every UPDATE_RATE milliseconds while playback is in progress.
  * @param onPlay optional function to run when playback starts.
  * @param onStop optional function to run when playback stops.
  */
 export default function usePlaybackUpdate(
+  isPlaying: boolean,
   doUpdate: () => void,
   onPlay?: () => void,
   onStop?: () => void
 ) {
-  const isPlaying = useAppSelector(state => state.music.isPlaying);
   const intervalId = useRef<number | undefined>();
 
   useEffect(() => {
